% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixNormal_pkg.R
\docType{package}
\name{matrixNormal}
\alias{matrixNormal}
\alias{matrixNormal-package}
\title{matrixNormal package}
\description{
A package for computing densities, probabilities and quantiles for The Matrix Normal Distribution. The matrixNormal package consists of two sections.
}
\section{Matrix functions}{

 Useful but simple matrix operations have been coded: \describe{
 \item{\code{is.symmetric.matrix}}{Is a matrix square? symmetric? positive definite? or positive semi-definite? A tolerance
 is included here.}
 \item{\code{Special_matrices}}{Creates the identity matrix I and matrix of 1's J.}
 \item{\code{tr}}{Calculates the usual trace of a matrix}
 \item{\code{vec}}{Stacks a Matrix using matrix operator "vec" and has option to keep names.}
 }
}

\section{Distribution functions}{

 This package also contains the PDF, CDF, and random number generation for the matrix normal
 distribution (*matnorm). See matrixNormal_Distribution.
}

\author{
: Paul M. Hargarten
}
