% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Matrix Trace}
\usage{
tr(A)
}
\arguments{
\item{A}{Square matrix.}
}
\description{
Computes the trace of a square numeric matrix A.
}
\note{
If the argument is not a square numeric matrix, the function presents an error and
     terminates.
}
\examples{
A <- matrix( seq( 1, 16, 1 ), nrow=4, byrow=TRUE )
A
tr( A )
tr( I(3) )

}
\seealso{
Other matrix: \code{\link{special.matrix}}
}
\concept{matrix}
\keyword{matrix}
